requirements.txt
setuptools (18.2)
pip (1.5.6)
six (1.9.0)
urllib3 (1.12)
MySQL-python (1.2.5)
DBUtils (1.1)
python-memcached (1.53)
pytz (2015.4)
python-gflages (2.0)
python-dateutil (2.4.2)
google-apputils (0.4.2)
protobuf (2.6.1)
configparser (3.5.0b2)
nose (1.3.3)
pycurl2 (7.20.0)
unittest2 (0.5.1)
virtualenv (1.11.6)
mockito (0.5.2)


